requirements.txt
fastapi==0.109.0
uvicorn==0.23.2
pydantic-settings==2.1.0
pillow==10.2.0
python-multipart==0.0.6
PyJWT==2.8.0